added withMerged to ChangelogCommand, which enables for listing merge commits in the changelogCommand output. Added unit tests to verify both the existing behavior and the new behavior
